requirements.txt
flask
vanna[chromadb,pgvector,openai,ollama,postgres,mysql]
db-dtypes
python-dotenv
langchain-huggingface
requests
psycopg2-binary
dashscope
psycopg[binary]